+++ app/Http/Controllers/GroupController.php
@@ -11,6 +11,7 @@ use App\Union;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class GroupController extends Controller
 {
@@ -46,43 +47,48 @@ class GroupController extends Controller
             }
         } catch (\Exception $e) {
             return response()->json([
-                "data" => null,
-            ], 404, HeaderHelper::$header);
+                "error" => $e,
+            ], 500, HeaderHelper::$header);
         }
         return response()->json([
             "data" => $data,
         ], 200, HeaderHelper::$header);
     }
 
-    // TODO delete this method (route api/groups/{id})
+    // TODO delete this method (route api/groups/{id}) after update client
     public function indexFromParams($user_id)
     {
-        if (Auth::user()->role_id == '1') {
-            $data = Group::with(['union' => function($query) {
+        try {
+            if (Auth::user()->role_id == '1') {
+                $data = Group::with(['union' => function($query) {
                     $query->select('id','name');
                 }])
-                ->with(['user' => function($query) {
-                    $query->select('id', 'email');
-                }])
-                ->orderBy('id', 'DESC')
-                ->get();
-        } else {
-            $union_id = Union::select('id')->where('user_id', Auth::id())->first();
-            $data = Group::select('id', 'name', 'user_id', 'union_id')
-                ->where('union_id', $union_id->id)
-                ->with(['union' => function($query) {
-                    $query->select('id','name');
-                }])
-                ->with(['user' => function($query) {
-                    $query->select('id', 'email');
-                }])
-                ->orderBy('id', 'DESC')
-                ->get();
+                    ->with(['user' => function($query) {
+                        $query->select('id', 'email');
+                    }])
+                    ->orderBy('id', 'DESC')
+                    ->get();
+            } else {
+                $union_id = Union::select('id')->where('user_id', Auth::id())->first();
+                $data = Group::select('id', 'name', 'user_id', 'union_id')
+                    ->where('union_id', $union_id->id)
+                    ->with(['union' => function($query) {
+                        $query->select('id','name');
+                    }])
+                    ->with(['user' => function($query) {
+                        $query->select('id', 'email');
+                    }])
+                    ->orderBy('id', 'DESC')
+                    ->get();
+            }
+        } catch (\Exception $e) {
+            return response()->json([
+                "error" => $e,
+            ], 500, HeaderHelper::$header);
         }
-
         return response()->json([
             "data" => $data,
-        ], 200);
+        ], 200, HeaderHelper::$header);
     }
 
     public function store(Request $request)
@@ -98,6 +104,17 @@ class GroupController extends Controller
         // dd($request->getClientIp());
         // dd($request->json());
         try {
+            $validator = Validator::make($request->all(),[
+                'name' => 'required|min:5|unique:users,name',
+                'email' => 'required|email|unique:users,email',
+                'password' => 'required|min:5',
+            ]);
+            if($validator->messages()->first()) {
+                return response()->json([
+                    "response" => $validator->messages()
+                ], 400 , HeaderHelper::$header);
+            }
+
             $code = GeneratorHelper::code('GROUP');
             $user_id = UserCRUD::create($request, $code, '4')->id;
             if (Auth::user()->role_id == '1') {
@@ -120,8 +137,8 @@ class GroupController extends Controller
         } catch (\Exception $e) {
             // TODO get more errors, example = the email is currently used
             return response()->json([
-                "response" => false
-            ], 200, HeaderHelper::$header);
+                "error" => $e,
+            ], 500, HeaderHelper::$header);
         }
         return response()->json([
             "response" => true
@@ -135,14 +152,26 @@ class GroupController extends Controller
          * The request contain:
          *  - id = group_id
          */
-        $data = Group::where('id', $id)
-            ->with(['union' => function($query) {
-                $query->select('id','name');
-            }])
-            ->with(['user' => function($query) {
-                $query->select('id', 'email');
-            }])
-            ->first();
+        try {
+            if (!Group::where('id', $id)->exists()) {
+                return response()->json([
+                    "error" => "The current id: $id does't  exists.",
+                ], 404, HeaderHelper::$header);
+            }
+            $data = Group::where('id', $id)
+                ->with(['union' => function($query) {
+                    $query->select('id','name');
+                }])
+                ->with(['user' => function($query) {
+                    $query->select('id', 'email');
+                }])
+                ->first();
+        } catch (\Exception $e) {
+            return response()->json([
+                "error" => $e,
+            ], 500, HeaderHelper::$header);
+        }
+
         return response()->json([
             "data" => $data,
         ], 200, HeaderHelper::$header);
@@ -163,8 +192,8 @@ class GroupController extends Controller
             UserCRUD::update($request);
         } catch (\Exception $e) {
             return response()->json([
-                'response' => false,
-            ],404, HeaderHelper::$header);
+                'error' => $e,
+            ],500, HeaderHelper::$header);
         }
         return response()->json([
             "response" => true,
@@ -177,6 +206,18 @@ class GroupController extends Controller
          * The request contain:
          *  - id = group_id
          */
-        UserCRUD::destroy($id);
+        try {
+            if (!$id->exists()) {
+                return response()->json([
+                    "error" => "The current id: $id does't  exists.",
+                ], 404, HeaderHelper::$header);
+            }
+            UserCRUD::destroy($id);
+        } catch (\Exception $e) {
+            return response()->json([
+                "error" => $e,
+            ], 500, HeaderHelper::$header);
+        }
+        return response()->json([],200, HeaderHelper::$header);
     }
 }
